Task #5081: Make sure buffer size calculation doesn't overflow

+++ branches/Cobalt-Task5081/RTCP/Cobalt/GPUProc/src/cuda/Kernels/IncoherentStokesTransposeKernel.cc
@@ -89,9 +89,10 @@ namespace LOFAR
       case IncoherentStokesTransposeKernel::INPUT_DATA:
       case IncoherentStokesTransposeKernel::OUTPUT_DATA:
         return 
-          itsParameters.nrStations * itsParameters.nrChannelsPerSubband *
-          itsParameters.nrSamplesPerChannel * NR_POLARIZATIONS *
-          sizeof(std::complex<float>);
+          (size_t) itsParameters.nrStations * 
+          itsParameters.nrChannelsPerSubband *
+          itsParameters.nrSamplesPerChannel * 
+          NR_POLARIZATIONS * sizeof(std::complex<float>);
       default:
         THROW(GPUProcException, "Invalid bufferType (" << bufferType << ")");
       }
